kortix-cli/requirements.txt
# Kortix CLI - 增强版依赖

# 核心 LLM
dashscope>=1.14.0        # 阿里云百炼

# Docker 沙箱
docker>=7.0.0

# CLI 界面
rich>=13.0.0
click>=8.0.0
prompt_toolkit>=3.0.0    # 高级CLI交互

# 配置
pyyaml>=6.0.0
python-dotenv>=1.0.0

# 日志
structlog>=25.0.0

# HTTP 请求
requests>=2.32.0
httpx>=0.27.0
aiohttp>=3.9.0

# Web 搜索 (可选)
tavily-python>=0.5.0     # Web搜索 API

# 图片处理
Pillow>=10.4.0

# 文件处理
PyPDF2>=3.0.0            # PDF 读取
python-docx>=1.1.0       # Word 文档
openpyxl>=3.1.0          # Excel
beautifulsoup4>=4.12.0   # HTML 解析
markdown>=3.5.0          # Markdown 处理

# 数据处理
pandas>=2.0.0            # 数据分析（可选但推荐）

